Added login form, greeting, logout button

+++ index.php
@@ -1,3 +1,17 @@
+<?php
+
+session_start();
+
+if (isset($_POST['logout'])) { // logout
+    session_destroy();
+    session_start();
+    unset($_SESSION['username']);
+    unset($_SESSION['password']);
+    unset($_SESSION['logged_in']);
+    header("Location: http://localhost/BIT_PP6/");
+}
+?>
+
 <!DOCTYPE html>
 <html lang="en">
 
@@ -12,7 +26,58 @@
 
 <body>
 
-    <?php
+    <?php // login
+    $msg = '';
+    if (isset($_POST['login']) && !empty($_POST['username']) && !empty($_POST['password'])) {
+        if ($_POST['username'] == 'admin' && $_POST['password'] == 'admin') {
+            $_SESSION['logged_in'] = true;
+            $_SESSION['timeout'] = time();
+            $_SESSION['username'] = 'Admin';
+            unset($_GET['logout']);
+        } else if ($_POST['username'] == 'user' && $_POST['password'] == 'user') {
+            $_SESSION['logged_in'] = true;
+            $_SESSION['timeout'] = time();
+            $_SESSION['username'] = 'User';
+        } else {
+            $msg = '<strong>Failed to login:</strong> wrong username and/or password!';
+        }
+    } else if (isset($_POST['logout'])) {
+        $msg = '<div class="loggedOutMsg"><strong>You successfully loged out</strong></div>';
+    }
+
+
+    if (isset($_SESSION['logged_in']) == false) {
+        echo ('<div class="loginFormWrap">
+                <div class="headerWrap">
+                    <h2>Log In info</h2>
+                    <p>Admin has more rights, User cannot delete files</p>
+                    <div class="loginInfo">
+                        <div class="loginAdmin">
+                            <p>Username: admin <br>
+                            Password: admin                            
+                            <br>
+                            </p>
+                        </div>
+                        <div class="loginUser">
+                            <p>Username: user <br>
+                            Password: user                            
+                            <br>
+                            </p>
+                        </div>
+                    </div>
+                </div>
+                <span class="errorMsg">' . $msg . '</span>
+                <form action = "" method = "post">
+                    <label for="username">Enter your username:<br></label>
+                    <input class="loginInput" type="text" id="username" name="username" required autofocus></br>
+                    <label for="password">Enter your password:<br></label>
+                    <input class="loginInput" type="password" id="password" name="password" required><br>
+                    <button class="loginButton" type="submit" name="login">Login</button>
+                </form>
+            </div>'
+        );
+        die();
+    }
 
     if (isset($_GET['path'])) {
         $path = './' . $_GET['path'];
@@ -79,7 +144,9 @@
     $downloadIcon = file_get_contents("./svg/download.svg");
 
     echo ("<div class='header'><button class='buttonBack'>" . "<a href='./?path=" . ltrim($emptyString, "./") . "'>" . "Back" . "</a>" . "</button>"); //Back button
-
+    echo "<div  class='user-greeting'><h1>Welcome," . ' ' . $_SESSION['username'] . '!' . "</h1></div>"; // Greeting
+    echo ("<form class='logout' action='' method='POST'><input type='submit' name='logout' value='Logout'></form></div>"); // Logout button
+    echo '<div class="pathContainer">Current directory:<span>' . ltrim($path, "./") . '/</span></div>'; //Current directory
 
     // Table start
     echo ("<table>
@@ -104,7 +171,11 @@
         }
         if (is_file($path . "/" . $item)) { // prevention to delete required files(styling and php)
             if ($item != "index.php") {
-                echo ("<td><a class='deleteButton'href='./?path=" . ltrim($path, "./") . "&file=" . $item . "&action=delete" . "'>" . "$deleteIcon</a><a class='downloadButton' href='./?path=" . ltrim($path, "./") . "&file=" . $item . "&action=download" . "'>" . "$downloadIcon</a></td>");
+                if ($_SESSION['username'] == 'Admin' && $item != "style.css" && $item != "buttons.css" && $item != "forms.css"  && $item != "delete.svg"  && $item != "download.svg") { // if ;admin can delete, user cannot.
+                    echo ("<td><a class='deleteButton'href='./?path=" . ltrim($path, "./") . "&file=" . $item . "&action=delete" . "'>" . "$deleteIcon</a><a class='downloadButton' href='./?path=" . ltrim($path, "./") . "&file=" . $item . "&action=download" . "'>" . "$downloadIcon</a></td>");
+                } else {
+                    echo ("<td><a class='downloadButton' href='./?path=" . ltrim($path, "./") . "&file=" . $item . "&action=download" . "'>" . "$downloadIcon</a></td>");
+                }
             } else {
                 echo ("<td></td>");
             }
